create_secrets/requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
google-auth==1.14.0
idna==2.9
kubernetes==11.0.0
mypy==0.770
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
websocket-client==0.57.0
